brush preview fix
The brush preview is now the accurate size all the time. The normal brush is still disabled so you have to choose one from the menu before painting. While the prieview changes size with the buttons, the actual brush does not, so we need to change the code to use the brush preview image to change the brush whenever it or the size buttons are clicked, insted of changing when you select a button from the menu. Also the colour change no longer works.