Dynamic gallery demo
designed a unique 3d space for the museum layout. move camera 5 (purple cylander) to the middle of the rooms and spin to get an idea of how it will look. explination of rooms are in my temp folder as a picture. using this 3d method we loose most of the menu ui and replace it with 3d spaces the player can visit.