Tutorial and New Museum Model
The tutorial is functioning and can be played without any bugs apart from the gameplay ones that were part of this version when i started which you guys have fixed. If part of the tutorial is off you may need to reset the player preffs to 0 in tutorial manager. The animations are a bit janky but i need to make changes to the ui layout before fixing them.
The first proper model for the gallery is set up in gallery scene although the scale needs to be changed to make lighting easier.