Navigation
we now have a developed gallery scene with models and scene navigation. the basis of the gallery navigation is also set up. i have remade all of matts models. the layout of scenes have been changed to scale correctly with the rest and to minimise ui. the transition system has been changed to be more efficient. animations and code have been added to alow for character reactions.